Refactored BtsLogAnalyzer module.

+++ BtsTraceDecoder/BtsTraceDecoder/src/BtsTraceDecoder.cpp
@@ -37,28 +37,28 @@ void BtsTraceDecoder::processInputTraceFile(std::string const& inputTraceFilePat
             unsigned int traceAddressValue = convertHexStringToNumber<unsigned int>(getStringInBetweenTwoStrings(lineInTraceFile, "[", "]"));
             if(traceAddressValue != 0)
             {
-                cout<<"TraceAddressValue: [0x"<<std::hex<<traceAddressValue<<"] NearestSymbol: ["<<getNearestLowerSymbol(traceAddressValue)<<"]"<<endl;
+                cout << "TraceAddressValue: [0x"
+                     << std::hex<<traceAddressValue
+                     << "] NearestSymbol: ["
+                     << getNearestLowerSymbol(static_cast<int>(traceAddressValue))
+                     << "]" << endl;
             }
         }
-    }
-}
+    }}
 
 std::string BtsTraceDecoder::getNearestLowerSymbol(int const address, int const offset)
-{
-    int addressWithOffset(address+offset);
+{    int addressWithOffset(address+offset);
     string symbol;
     if(!m_symbolMap.empty())
     {
         BtsTraceDecoder::SymbolMapType::iterator symbolIterator = m_symbolMap.lower_bound(addressWithOffset);
-        if(symbolIterator->first <= addressWithOffset)
+        if(static_cast<int>(symbolIterator->first) <= addressWithOffset)
         {
             symbol = symbolIterator->second;
-        }
-        else
+        }        else
         {
             if(symbolIterator != m_symbolMap.begin())
-            {
-                symbolIterator--;
+            {                symbolIterator--;
                 symbol = symbolIterator->second;
             }
         }
